refactor(openomni): split ingress handlers
